requirements.txt
rembg
onnxruntime
Pillow
gradio>=4.0.0 